+++ test/test_cl/test_operators/test_nft.py
@@ -13,6 +13,7 @@
 #
 # Copyright(C) 2019-2021 Max-Planck-Society
 # Copyright(C) 2022 Max-Planck-Society, Philipp Arras
+# Copyright(C) 2025 LambdaFields GmbH
 #
 # NIFTy is being developed at the Max-Planck-Institut fuer Astrophysik.
 
@@ -203,3 +204,110 @@ def test_build_nufft(space, N, eps):
     cmplx = np.complex128
     ift.extra.check_linear_operator(RF, cmplx, flt, only_r_linear=True,
                                     rtol=eps, atol=eps, no_device_copies=False)
+
+
+@pmp('eps', [1e-4, 1e-7, 1e-10, 1e-11, 1e-12, 2e-13])
+@pmp('shape', [(34,), (127,), (34, 50), (129, 100), (10, 10, 10)])
+@pmp('N', [1, 10, 100])
+@pmp('pre_domain', [None, ift.UnstructuredDomain(3)])
+def test_variable_position_nufft_consistency(shape, N, eps, pre_domain):
+    dst = (0.2, 0.99, 9.213)
+    ndim = len(shape)
+    dom = ift.RGSpace(shape, dst[:ndim])
+    dst = np.array(dom.distances)
+
+    ddom = ift.UnstructuredDomain(N), ift.UnstructuredDomain(ndim)
+    coord = (ift.from_random(ddom).asnumpy() - 0.5) / dst[None]
+    coord = ift.makeField(ddom, coord)
+
+    op = ift.VariablePositionNufft(dom, N, epsilon=eps, pre_domain=pre_domain)
+    inp = ift.MultiField.from_dict({"grid": ift.from_random(op.domain["grid"], dtype=complex),
+                                    "coord": coord})
+    res = op(inp).asnumpy()
+
+    ref = np.empty_like(res)
+    grid = inp["grid"].asnumpy()
+    coord = inp["coord"].asnumpy()
+
+    if pre_domain is None:
+        grid = grid[None]
+        ref = ref[None]
+    corfact = -2j*np.pi*dst
+    xyz = np.meshgrid(*[(-(ss//2) + np.arange(ss))*cf
+                        for ss, cf in zip(dom.shape, corfact)],
+                      indexing='ij')
+    for islc in range(grid.shape[0]):
+        for i in range(N):
+            phase = sum([a*b for a,b in zip(xyz, coord[i])])
+            ref[islc, i] = np.sum(grid[islc]*np.exp(phase))
+
+    if pre_domain is None:
+        ref = ref[0]
+
+    ift.myassert(_l2error(ref, res) < eps)
+    ift.extra.check_operator(op, inp, ntries=5, tol=10*eps)
+
+
+@pmp('eps', [1e-4, 1e-7, 1e-10, 1e-11, 1e-12, 2e-13])
+@pmp('shape', [(34,), (1001,), (100, 100), (35, 199), (20, 20, 21)])
+@pmp('pre_domain', [None, ift.UnstructuredDomain(3)])
+def test_variable_position_nufft_vs_fft(shape, eps, pre_domain):
+    dst = (0.2, 0.99, 28.1)
+    ndim = len(shape)
+    dom = ift.RGSpace(shape, dst[:ndim])
+    if pre_domain is not None:
+        dom = (pre_domain, dom)
+    dom = ift.DomainTuple.make(dom)
+
+    op1 = ift.ShiftedPositionFFT(dom[-1], eps, pre_domain)
+    op0 = ift.FFTOperator(dom, space=None if pre_domain is None else 1)
+    assert op1.domain["grid"] == ift.makeDomain(dom)
+    assert op1.domain["delta_coord"] == ift.makeDomain((dom[-1], ift.UnstructuredDomain(ndim)))
+    inp = ift.MultiField.from_dict({"grid": ift.from_random(op1.domain["grid"], dtype=complex)},
+                                   domain=op1.domain)
+    ref = op0(inp["grid"]).asnumpy()
+    res = op1(inp).asnumpy()
+    ift.myassert(_l2error(ref, res) < eps)
+
+    delta_coord = np.zeros(op1.domain["delta_coord"].shape)
+    if ndim == 1:
+        delta_coord[1, 0] = -1
+        delta_coord[2, 0] = 2
+    elif ndim == 2:
+        delta_coord[1, 0, 0] = -1
+        delta_coord[0, 2, 1] = 2
+        delta_coord[3, 3, 0] = 1
+        delta_coord[3, 3, 1] = -2
+    elif ndim == 3:
+        delta_coord[6, 4, 3, 0] = 3
+        delta_coord[6, 4, 3, 1] = -1
+        delta_coord[6, 4, 3, 2] = -2
+        delta_coord[2, 1, 10, 0] = -5
+        delta_coord[2, 1, 10, 1] = 20
+        delta_coord[2, 1, 10, 2] = 30
+
+    delta_coord = ift.makeField(op1.domain["delta_coord"], delta_coord)
+    inp = ift.MultiField.from_dict(
+        {
+            "grid": ift.from_random(dom, dtype=complex),
+            "delta_coord": delta_coord,
+        }
+    )
+    res = op1(inp).asnumpy()
+    if pre_domain is None:
+        res = res[None]
+        ntrans = 1
+    else:
+        ntrans = pre_domain.shape[0]
+
+    for itrans in range(ntrans):
+        if ndim == 1:
+            np.testing.assert_allclose(res[itrans, 0], res[itrans, 1])
+            np.testing.assert_allclose(res[itrans, 2], res[itrans, 4])
+        elif ndim == 2:
+            np.testing.assert_allclose(res[itrans, 1, 0], res[itrans, 0, 0])
+            np.testing.assert_allclose(res[itrans, 0, 2], res[itrans, 0, 4])
+            np.testing.assert_allclose(res[itrans, 3, 3], res[itrans, 4, 1])
+        elif ndim == 3:
+            np.testing.assert_allclose(res[itrans, 6, 4, 3], res[itrans, 9, 3, 1])
+            np.testing.assert_allclose(res[itrans, 2, 1, 10], res[itrans, 17, 1, 19]) # Test wraparound